refactor: 폴더구조 정리(#106)
- stores/auth 중복 폴더 삭제
- store → stores 폴더명 변경
- features/exams/questions → features/exams/exam-managements/questions 이동
