+++ src/features/exams/index.ts
@@ -5,14 +5,14 @@ export { mockExamList } from '../../mocks/mockExamList.ts'
 export { default as ExamCreateModal } from './components/ExamCreateModal.tsx'
 export { createExamDropdown } from './utils/createExamDropdown.ts'
 export type { ExamQuestion, ExamQuestionResponse } from './types.ts'
-export { ExamQuestionDetailBody } from '../exams/questions/components/detail-modal/ExamQuestionDetailBody.tsx'
-export { ExamQuestionDetailFooter } from '../exams/questions/components/detail-modal/ExamQuestionDetailFooter.tsx'
-export { ExamQuestionDetailSide } from '../exams/questions/components/detail-modal/ExamQuestionDetailSide.tsx'
-export { default as ExamQuestionDetailModal } from '../exams/questions/components/detail-modal/ExamQuestionDetailModal.tsx'
-export { default as ExamDeploymentsModal } from '../exams/components/ExamDeploymentsModal.tsx'
+export { ExamQuestionDetailBody } from './exam-managements/questions//detail-modal/ExamQuestionDetailBody.tsx'
+export { ExamQuestionDetailFooter } from './exam-managements/questions/detail-modal/ExamQuestionDetailFooter.tsx'
+export { ExamQuestionDetailSide } from './exam-managements/questions/detail-modal/ExamQuestionDetailSide.tsx'
+export { default as ExamQuestionDetailModal } from './exam-managements/questions/detail-modal/ExamQuestionDetailModal.tsx'
+export { default as ExamDeploymentsModal } from './components/ExamDeploymentsModal.tsx'
 export { default as ExamPage } from '../../pages/ExamManagementPage.tsx'
-export * from '../exams/questions/types.ts'
-export { default as ExamDeletePopupModal } from '../exams/questions/components/detail-modal/ExamDeletePopupModal.tsx'
+export * from './exam-managements/questions/types.ts'
+export { default as ExamDeletePopupModal } from './exam-managements/questions/detail-modal/ExamDeletePopupModal.tsx'
 export { default as DistributionList } from './deployments/components/DistributionList.tsx'
 export { default as SubmissionList } from './submissions/components/SubmissionList.tsx'
 export type { Exam, Distribution, Submission } from './types.ts'
